+++ CHANGELOG.md
@@ -1,7 +1,18 @@
-## 0.1.1
 
-- Fixed description.
+# 0.1.3
 
-## 0.1.0
+* **[IMPROVEMENT]** Fixed Readme not showing the right status
+* **[FEATURE]** Added unit tests
+* **[FEATURE]** Added row header names/index
 
-- Initial version.
+# 0.1.2
+
+* Fixed description to match dart packaging
+
+# 0.1.1
+
+* Fixed description.
+
+# 0.1.0
+
+* Initial version.